+++ components/src/stories/PromptModal.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import PromptModal from './PromptModal';
+import { PromptModal } from '../lib';
 
 const meta: Meta<typeof PromptModal> = {
   title: 'Components/PromptModal',
@@ -16,6 +16,10 @@ export const PromptModalModalCenter: Story = {
     isOpen: true,
     position: 'center',
     size: 'small',
+    title: '타이틀',
+    value: '',
+    onChange: () => {},
+    onSubmit: () => {},
     onClose: () => {},
     onBackdropClick: () => {},
   },
@@ -25,6 +29,10 @@ export const PromptModalModalBottom: Story = {
   args: {
     isOpen: true,
     position: 'bottom',
+    title: '타이틀',
+    value: '',
+    onChange: () => {},
+    onSubmit: () => {},
     onClose: () => {},
     onBackdropClick: () => {},
   },
@@ -35,6 +43,10 @@ export const PromptModalModalSmall: Story = {
     isOpen: true,
     position: 'center',
     size: 'small',
+    title: '타이틀',
+    value: '',
+    onChange: () => {},
+    onSubmit: () => {},
     onClose: () => {},
     onBackdropClick: () => {},
   },
@@ -45,6 +57,10 @@ export const PromptModalMedium: Story = {
     isOpen: true,
     position: 'center',
     size: 'medium',
+    title: '타이틀',
+    value: '',
+    onChange: () => {},
+    onSubmit: () => {},
     onClose: () => {},
     onBackdropClick: () => {},
   },
@@ -55,6 +71,10 @@ export const PromptModalLarge: Story = {
     isOpen: true,
     position: 'center',
     size: 'large',
+    title: '타이틀',
+    value: '',
+    onChange: () => {},
+    onSubmit: () => {},
     onClose: () => {},
     onBackdropClick: () => {},
   },
